Account Number String refactor
