+++ src/main/java/com/bae/entities/UserNotFoundException.java
@@ -1,5 +1,11 @@
 package com.bae.entities;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class UserNotFoundException extends RuntimeException {
+
+	private static final long serialVersionUID = 4767136409898040810L;
 
 }
