+++ src/main/java/com/bae/entities/UserRepo.java
@@ -6,4 +6,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepo extends JpaRepository<User, Long> {
 
+	boolean existsByAccountNumber(String accountNumber);
+
 }
